fix task wallet

+++ UserManager/views.py
@@ -1075,12 +1075,13 @@ def change_username(request):
         })
     request.user.username= username
     request.user.fullname = fullname
-    if avatar_id != request.user.avatar.id:
-        request.user.avatar.delete()
-        main_file = MainFile.objects.get(id=avatar_id)
-        main_file.its_blong = True
-        main_file.save()
-        request.user.avatar = main_file
+    if avatar_id:
+        if avatar_id != request.user.avatar.id:
+            request.user.avatar.delete()
+            main_file = MainFile.objects.get(id=avatar_id)
+            main_file.its_blong = True
+            main_file.save()
+            request.user.avatar = main_file
     request.user.save()
     return Response(status=status.HTTP_200_OK,data={
             "status":True,
